split up into "core" & "redux-hooks" module (#116)
