+++ examples/svelte/src/main.ts
@@ -1,10 +1,12 @@
 import App from './App.svelte';
 import { worker } from './mocks/browser';
 
-worker.start();
-
-const app = new App({
-    target: document.body,
-});
+// Defer the load to ensure the worker has started.
+const app = worker.start({ onUnhandledRequest: 'error' }).then(
+    () =>
+        new App({
+            target: document.body,
+        }),
+);
 
 export default app;
