+++ src/apiTypes.ts
@@ -1,106 +1,54 @@
-/**
- * Note: this file should import all other files for type discovery and declaration merging
- */
-import { PatchQueryResultThunk, QueryApi, UpdateQueryResultThunk } from './buildThunks';
-import { AnyAction, Middleware, Reducer, ThunkDispatch, ThunkAction } from '@reduxjs/toolkit';
-import { PrefetchOptions } from './buildHooks';
-import {
-  EndpointDefinitions,
-  EndpointBuilder,
-  QueryArgFrom,
-  QueryDefinition,
-  MutationDefinition,
-} from './endpointDefinitions';
-import { CombinedState, QueryKeys, RootState } from './apiState';
-import { UnionToIntersection } from './tsHelpers';
-import { TS41Hooks } from './ts41Types';
-import './buildSelectors';
-import { SliceActions } from './buildSlice';
-import { onFocus, onFocusLost, onOffline, onOnline } from './setupListeners';
+import { EndpointDefinitions, EndpointBuilder, EndpointDefinition } from './endpointDefinitions';
+import { UnionToIntersection, Id } from './tsHelpers';
+import { CoreModule } from './core/module';
+import { CreateApiOptions } from './createApi';
+import { BaseQueryFn } from './baseQueryTypes';
 
-type UnwrapPromise<T> = T extends PromiseLike<infer V> ? V : T;
-type MaybePromise<T> = T | PromiseLike<T>;
+export interface ApiModules<
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  BaseQuery extends BaseQueryFn,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  Definitions extends EndpointDefinitions,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  ReducerPath extends string,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  EntityTypes extends string
+> {}
 
-export type BaseQueryFn<Args = any, Result = unknown, Error = unknown, DefinitionExtraOptions = {}> = (
-  args: Args,
-  api: QueryApi,
-  extraOptions: DefinitionExtraOptions
-) => MaybePromise<{ error: Error; data?: undefined } | { error?: undefined; data?: Result }>;
+export type ModuleName = keyof ApiModules<any, any, any, any>;
 
-export type BaseQueryEnhancer<AdditionalArgs = unknown, AdditionalDefinitionExtraOptions = unknown, Config = void> = <
-  BaseQuery extends BaseQueryFn
->(
-  baseQuery: BaseQuery,
-  config: Config
-) => BaseQueryFn<
-  BaseQueryArg<BaseQuery> & AdditionalArgs,
-  BaseQueryResult<BaseQuery>,
-  BaseQueryError<BaseQuery>,
-  BaseQueryExtraOptions<BaseQuery> & AdditionalDefinitionExtraOptions
->;
-
-export type BaseQueryResult<BaseQuery extends BaseQueryFn> = Exclude<
-  UnwrapPromise<ReturnType<BaseQuery>>,
-  { data: undefined }
->['data'];
-
-export type BaseQueryError<BaseQuery extends BaseQueryFn> = Exclude<
-  UnwrapPromise<ReturnType<BaseQuery>>,
-  { error: undefined }
->['error'];
-
-export type BaseQueryArg<T extends (arg: any, ...args: any[]) => any> = T extends (arg: infer A, ...args: any[]) => any
-  ? A
-  : any;
-
-export type BaseQueryExtraOptions<BaseQuery extends BaseQueryFn> = Parameters<BaseQuery>[2];
+export type Module<Name extends ModuleName> = {
+  name: Name;
+  init<
+    BaseQuery extends BaseQueryFn,
+    Definitions extends EndpointDefinitions,
+    ReducerPath extends string,
+    EntityTypes extends string
+  >(
+    api: Api<BaseQuery, EndpointDefinitions, ReducerPath, EntityTypes, ModuleName>,
+    options: Required<CreateApiOptions<BaseQuery, Definitions, ReducerPath, EntityTypes>>,
+    context: {
+      endpointDefinitions: Definitions;
+    }
+  ): {
+    injectEndpoint(endpoint: string, definition: EndpointDefinition<any, any, any, any>): void;
+  };
+};
 
 export type Api<
   BaseQuery extends BaseQueryFn,
   Definitions extends EndpointDefinitions,
   ReducerPath extends string,
-  EntityTypes extends string
-> = {
-  reducerPath: ReducerPath;
-  internalActions: InternalActions;
-  reducer: Reducer<CombinedState<Definitions, EntityTypes, ReducerPath>, AnyAction>;
-  middleware: Middleware<{}, RootState<Definitions, string, ReducerPath>, ThunkDispatch<any, any, AnyAction>>;
-  util: {
-    updateQueryResult: UpdateQueryResultThunk<Definitions, RootState<Definitions, string, ReducerPath>>;
-    patchQueryResult: PatchQueryResultThunk<Definitions, RootState<Definitions, string, ReducerPath>>;
-  };
-  // If you actually care about the return value, use useQuery
-  usePrefetch<EndpointName extends QueryKeys<Definitions>>(
-    endpointName: EndpointName,
-    options?: PrefetchOptions
-  ): (arg: QueryArgFrom<Definitions[EndpointName]>, options?: PrefetchOptions) => void;
-  injectEndpoints<NewDefinitions extends EndpointDefinitions>(_: {
-    endpoints: (build: EndpointBuilder<BaseQuery, EntityTypes, ReducerPath>) => NewDefinitions;
-    overrideExisting?: boolean;
-  }): Api<BaseQuery, Definitions & NewDefinitions, ReducerPath, EntityTypes>;
-  endpoints: {
-    [K in keyof Definitions]: Definitions[K] extends QueryDefinition<any, any, any, any, any>
-      ? ApiEndpointQuery<Definitions[K], Definitions>
-      : Definitions[K] extends MutationDefinition<any, any, any, any, any>
-      ? ApiEndpointMutation<Definitions[K], Definitions>
-      : never;
-  };
-} & TS41Hooks<Definitions>;
-
-export interface ApiEndpointQuery<
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  Definition extends QueryDefinition<any, any, any, any, any>,
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  Definitions extends EndpointDefinitions
-> {}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export interface ApiEndpointMutation<
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  Definition extends MutationDefinition<any, any, any, any, any>,
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  Definitions extends EndpointDefinitions
-> {}
+  EntityTypes extends string,
+  Enhancers extends ModuleName = CoreModule
+> = Id<
+  Id<UnionToIntersection<ApiModules<BaseQuery, Definitions, ReducerPath, EntityTypes>[Enhancers]>> & {
+    injectEndpoints<NewDefinitions extends EndpointDefinitions>(_: {
+      endpoints: (build: EndpointBuilder<BaseQuery, EntityTypes, ReducerPath>) => NewDefinitions;
+      overrideExisting?: boolean;
+    }): Api<BaseQuery, Definitions & NewDefinitions, ReducerPath, EntityTypes, Enhancers>;
+  }
+>;
 
 export type ApiWithInjectedEndpoints<
   ApiDefinition extends Api<any, any, any, any>,
@@ -111,20 +59,3 @@ export type ApiWithInjectedEndpoints<
   Omit<Injections, 'endpoints'> & {
     endpoints: ApiDefinition['endpoints'] & Partial<UnionToIntersection<Injections[number]['endpoints']>>;
   };
-
-export type InternalActions = SliceActions & {
-  prefetchThunk: (endpointName: any, arg: any, options: PrefetchOptions) => ThunkAction<void, any, any, AnyAction>;
-} & {
-  /**
-   * Will cause the RTK Query middleware to trigger any refetchOnReconnect-related behavior
-   * @link https://rtk-query-docs.netlify.app/api/setupListeners
-   */
-  onOnline: typeof onOnline;
-  onOffline: typeof onOffline;
-  /**
-   * Will cause the RTK Query middleware to trigger any refetchOnFocus-related behavior
-   * @link https://rtk-query-docs.netlify.app/api/setupListeners
-   */
-  onFocus: typeof onFocus;
-  onFocusLost: typeof onFocusLost;
-};
