+++ src/core/index.ts
@@ -1,0 +1,6 @@
+import { buildCreateApi } from '../createApi';
+import { coreModule } from './module';
+
+const createApi = buildCreateApi(coreModule);
+
+export { createApi, coreModule };
