+++ src/react-hooks/index.ts
@@ -1,0 +1,7 @@
+import { coreModule } from '../core/module';
+import { buildCreateApi } from '../createApi';
+import { reactHooksModule } from './module';
+
+const createApi = buildCreateApi(coreModule, reactHooksModule);
+
+export { createApi, reactHooksModule };
