+++ src/retry.ts
@@ -1,5 +1,5 @@
-import { BaseQueryEnhancer } from './apiTypes';
-import { HandledError } from './buildThunks';
+import { BaseQueryEnhancer } from './baseQueryTypes';
+import { HandledError } from './core/buildThunks';
 
 async function defaultBackoff(attempt: number = 0, maxRetries: number = 5) {
   const attempts = Math.min(attempt, maxRetries);
